handle serialization_alias in pydantic 1
